migrated all the static callback classes to their own classes in the automated package
